refactor: Unit Tests - Querying All Users Registered in the Database.
Using two options:

One: Use the Persistence Entity Manager to query a user registered in the database,
handling the return through a persistence query interface.

Second: Use the Persistence Entity Manager to query a user registered in the database,
handling the return through the object list interface of the Java api.