refactor: Unit Tests - Update a Registered User in the Database.
Using three options:

- First: Use the Persistence Entity Manager to query a user registered in the database in one instance.
This overwrites the name and email address fields, and returns a request to the database to merge the data.

- Second: Use the Persistence Entity Manager to query a user registered in the database on one instance.
In it, you overwrite the name and email address fields, and return to the database a data allocation request
pointing to the same primary key already registered.

- Third: Use the Persistence Entity Manager to query a user registered in the database in one instance.
In it, you overwrite the name and email address fields. A request is sent to the database to highlight
the existing registration. The instance containing the overwritten parameters is sent to the database,
with the resuisition of being allocated to the same record that was highlighted.